Added disable verbose. Moved utils into utils.py.
