+++ src/containerapp/azext_containerapp/_utils.py
@@ -5,7 +5,6 @@
 # pylint: disable=line-too-long, consider-using-f-string, no-else-return, duplicate-string-formatting-argument
 
 from urllib.parse import urlparse
-from azure.cli.command_modules.appservice.custom import (_get_acr_cred)
 from azure.cli.core.azclierror import (ValidationError, RequiredArgumentMissingError)
 from azure.cli.core.commands.client_factory import get_subscription_id
 from knack.log import get_logger
@@ -585,3 +584,197 @@ def _registry_exists(containerapp_def, registry_server):
                 exists = True
                 break
     return exists
+
+
+def get_randomized_name(prefix, name=None, initial="rg"):
+    from random import randint
+    default = "{}_{}_{:04}".format(prefix, initial, randint(0, 9999))
+    if name is not None:
+        return name
+    return default
+
+
+def _set_webapp_up_default_args(cmd, resource_group_name, location, name, registry_server):
+    from azure.cli.core.util import ConfiguredDefaultSetter
+    with ConfiguredDefaultSetter(cmd.cli_ctx.config, True):
+        logger.warning("Setting 'az containerapp up' default arguments for current directory. "
+                       "Manage defaults with 'az configure --scope local'")
+
+
+        cmd.cli_ctx.config.set_value('defaults', 'resource_group_name', resource_group_name)
+        logger.warning("--resource-group/-g default: %s", resource_group_name)
+
+
+        cmd.cli_ctx.config.set_value('defaults', 'location', location)
+        logger.warning("--location/-l default: %s", location)
+
+
+        cmd.cli_ctx.config.set_value('defaults', 'name', name)
+        logger.warning("--name/-n default: %s", name)
+
+        # cmd.cli_ctx.config.set_value('defaults', 'managed_env', managed_env)
+        # logger.warning("--environment default: %s", managed_env)
+
+        cmd.cli_ctx.config.set_value('defaults', 'registry_server', registry_server)
+        logger.warning("--registry-server default: %s", registry_server)
+
+
+def get_profile_username():
+    from azure.cli.core._profile import Profile
+    user = Profile().get_current_account_user()
+    user = user.split('@', 1)[0]
+    if len(user.split('#', 1)) > 1:  # on cloudShell user is in format live.com#user@domain.com
+        user = user.split('#', 1)[1]
+    return user
+
+
+def create_resource_group(cmd, rg_name, location):
+    from azure.cli.core.profiles import ResourceType, get_sdk
+    rcf = _resource_client_factory(cmd.cli_ctx)
+    resource_group = get_sdk(cmd.cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES, 'ResourceGroup', mod='models')
+    rg_params = resource_group(location=location)
+    return rcf.resource_groups.create_or_update(rg_name, rg_params)
+
+
+def get_resource_group(cmd, rg_name):
+    from azure.cli.core.profiles import ResourceType, get_sdk
+    rcf = _resource_client_factory(cmd.cli_ctx)
+    resource_group = get_sdk(cmd.cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES, 'ResourceGroup', mod='models')
+    return rcf.resource_groups.get(rg_name)
+
+
+def _resource_client_factory(cli_ctx, **_):
+    from azure.cli.core.commands.client_factory import get_mgmt_service_client
+    from azure.cli.core.profiles import ResourceType
+    return get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
+
+
+def queue_acr_build(cmd, registry_rg, registry_name, img_name, src_dir, dockerfile="Dockerfile", disable_verbose=False):
+    import os, uuid, tempfile
+    from azure.cli.command_modules.acr._archive_utils import upload_source_code
+    from azure.cli.command_modules.acr._stream_utils import stream_logs
+    from azure.cli.command_modules.acr._client_factory import cf_acr_registries_tasks
+    from azure.cli.core.commands import LongRunningOperation
+
+    # client_registries = get_acr_service_client(cmd.cli_ctx).registries
+    client_registries = cf_acr_registries_tasks(cmd.cli_ctx)
+
+
+    if not os.path.isdir(src_dir):
+        raise CLIError("Source directory should be a local directory path.")
+
+
+    docker_file_path = os.path.join(src_dir, dockerfile)
+    if not os.path.isfile(docker_file_path):
+        raise CLIError("Unable to find '{}'.".format(docker_file_path))
+
+
+    # NOTE: os.path.basename is unable to parse "\" in the file path
+    original_docker_file_name = os.path.basename(docker_file_path.replace("\\", "/"))
+    docker_file_in_tar = '{}_{}'.format(uuid.uuid4().hex, original_docker_file_name)
+    tar_file_path = os.path.join(tempfile.gettempdir(), 'build_archive_{}.tar.gz'.format(uuid.uuid4().hex))
+
+
+    source_location = upload_source_code(cmd, client_registries, registry_name, registry_rg, src_dir, tar_file_path, docker_file_path, docker_file_in_tar)
+
+    # For local source, the docker file is added separately into tar as the new file name (docker_file_in_tar)
+    # So we need to update the docker_file_path
+    docker_file_path = docker_file_in_tar
+
+
+    from azure.cli.core.profiles import ResourceType
+    OS, Architecture = cmd.get_models('OS', 'Architecture', resource_type=ResourceType.MGMT_CONTAINERREGISTRY, operation_group='runs')
+    # Default platform values
+    platform_os = OS.linux.value
+    platform_arch = Architecture.amd64.value
+    platform_variant = None
+
+
+    DockerBuildRequest, PlatformProperties = cmd.get_models('DockerBuildRequest', 'PlatformProperties',
+                                                            resource_type=ResourceType.MGMT_CONTAINERREGISTRY, operation_group='runs')
+    docker_build_request = DockerBuildRequest(
+        image_names=[img_name],
+        is_push_enabled=True,
+        source_location=source_location,
+        platform=PlatformProperties(
+            os=platform_os,
+            architecture=platform_arch,
+            variant=platform_variant
+        ),
+        docker_file_path=docker_file_path,
+        timeout=None,
+        arguments=[])
+
+
+    queued_build = LongRunningOperation(cmd.cli_ctx)(client_registries.begin_schedule_run(
+        resource_group_name=registry_rg,
+        registry_name=registry_name,
+        run_request=docker_build_request))
+
+
+    run_id = queued_build.run_id
+    logger.warning("Queued a build with ID: %s", run_id)
+    not disable_verbose and logger.warning("Waiting for agent...")
+
+    from azure.cli.command_modules.acr._client_factory import (cf_acr_runs)
+    client_runs = cf_acr_runs(cmd.cli_ctx)
+
+    if disable_verbose:
+        logger.warning("Waiting for build to finish.")
+        finished = False
+        while not finished:
+            try: 
+                res = client_runs.get(registry_rg, registry_name, run_id)
+            except:
+                pass
+            if res.status != "Running":
+                if res.status != "Succeeded":
+                    logger.warning("Build failed.")
+                    return res.status
+                finished = True  # doesn't matter can just do while True
+                logger.warning("Build succeeded.")
+                return res.status
+
+        return client_runs.get(registry_rg, registry_name, run_id)  # returns only the response object
+
+    return stream_logs(cmd, client_runs, run_id, registry_name, registry_rg, None, False, True)
+
+
+def _get_acr_cred(cli_ctx, registry_name):
+    from azure.mgmt.containerregistry import ContainerRegistryManagementClient
+    from azure.cli.core.commands.parameters import get_resources_in_subscription
+    from azure.cli.core.commands.client_factory import get_mgmt_service_client
+
+    client = get_mgmt_service_client(cli_ctx, ContainerRegistryManagementClient).registries
+
+    result = get_resources_in_subscription(cli_ctx, 'Microsoft.ContainerRegistry/registries')
+    result = [item for item in result if item.name.lower() == registry_name]
+    if not result or len(result) > 1:
+        raise ResourceNotFoundError("No resource or more than one were found with name '{}'.".format(registry_name))
+    resource_group_name = parse_resource_id(result[0].id)['resource_group']
+
+    registry = client.get(resource_group_name, registry_name)
+
+    if registry.admin_user_enabled:  # pylint: disable=no-member
+        cred = client.list_credentials(resource_group_name, registry_name)
+        return cred.username, cred.passwords[0].value, resource_group_name
+    raise ResourceNotFoundError("Failed to retrieve container registry credentials. Please either provide the "
+                                "credentials or run 'az acr update -n {} --admin-enabled true' to enable "
+                                "admin first.".format(registry_name))
+
+
+def create_new_acr(cmd, registry_name, resource_group_name, location=None, sku="Basic"):
+    # from azure.cli.command_modules.acr.custom import acr_create
+    from azure.cli.command_modules.acr._client_factory import cf_acr_registries
+    from azure.cli.core.profiles import ResourceType
+
+
+    client = cf_acr_registries(cmd.cli_ctx)
+    # return acr_create(cmd, client, registry_name, resource_group_name, sku, location)
+    
+    Registry, Sku, NetworkRuleSet = cmd.get_models('Registry', 'Sku', 'NetworkRuleSet', resource_type=ResourceType.MGMT_CONTAINERREGISTRY, operation_group="registries")
+    registry = Registry(location=location, sku=Sku(name=sku), admin_user_enabled=True,
+                        zone_redundancy=None, tags=None)
+
+    # lro_poller = client.begin_create(resource_group_name, registry_name, registry, polling=False)
+    return client._create_initial(resource_group_name, registry_name, registry)
